Continue refactoring.
